+++ src/component/library/library.tsx
@@ -10,6 +10,7 @@ import { bookGroupingOptions } from './book-grouping-options'
 import { Icon } from '../icon/icon'
 import { SlidersHorizontal } from 'lucide-react'
 import { Button } from '../button/button'
+import { Link } from '../../bookshelf/book/link'
 
 type ViewType = 'gallery' | 'table'
 
@@ -83,7 +84,15 @@ export function Library({ settings, settingsChanged, books, sortOptions, onBookC
                 return false
             }
 
-            return b.metadata.title.toLowerCase().includes(settings.search.toLowerCase())
+            const searchableFields = [b.metadata.title.toLowerCase()]
+            const series = b.metadata.series
+            if (series) {
+                searchableFields.push(series.name instanceof Link ? series.name.displayText : series.name)
+            }
+
+            searchableFields.push(...b.metadata.authors.map((a) => (a instanceof Link ? a.displayText : a)))
+
+            return searchableFields.some((f) => f.toLowerCase().includes(settings.search.toLowerCase()))
         })
         .sort(sortOption?.compareFn)