+++ src/bookshelf/bookshelf.spec.ts
@@ -29,6 +29,8 @@ const defaultConfiguration: Configuration = {
             lists: 'lists',
             comment: 'comment',
             links: 'links',
+            series: 'series',
+            positionInSeries: 'position-in-series',
         },
         bookNote: {
             enabled: true,