+++ src/obsidian/settings/bookshelf-plugin-settings.ts
@@ -21,6 +21,8 @@ export interface BookshelfPluginSettings extends VersionedSettings {
         lists: string
         comment: string
         links: string
+        series: string
+        positionInSeries: string
     }
     bookNote: {
         enabled: boolean
@@ -65,6 +67,8 @@ export const DEFAULT_SETTINGS: Partial<BookshelfPluginSettings> = {
         lists: 'lists',
         comment: 'comment',
         links: 'links',
+        series: 'series',
+        positionInSeries: 'position-in-series',
     },
     bookNote: {
         enabled: true,