+++ CHANGELOG.md
@@ -7,7 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-For a full diff see [`0.13.1...main`][0.13.1...main].
+For a full diff see [`0.14.0...main`][0.14.0...main].
+
+## [`0.14.0`][0.14.0] - 2025-05-05
+
+For a full diff see [`0.13.1...0.14.0`][0.13.1...0.14.0].
 
 ### Added
 
@@ -252,6 +256,7 @@ For a full diff see [`0.2.2...0.3.0`][0.2.2...0.3.0].
 [0.12.0]: https://github.com/weph/obsidian-bookshelf/releases/tag/0.12.0
 [0.13.0]: https://github.com/weph/obsidian-bookshelf/releases/tag/0.13.0
 [0.13.1]: https://github.com/weph/obsidian-bookshelf/releases/tag/0.13.1
+[0.14.0]: https://github.com/weph/obsidian-bookshelf/releases/tag/0.14.0
 [0.2.2...0.3.0]: https://github.com/weph/obsidian-bookshelf/compare/0.2.2...0.3.0
 [0.3.0...0.3.1]: https://github.com/weph/obsidian-bookshelf/compare/0.3.0...0.3.1
 [0.3.1...0.4.0]: https://github.com/weph/obsidian-bookshelf/compare/0.3.1...0.4.0
@@ -266,4 +271,5 @@ For a full diff see [`0.2.2...0.3.0`][0.2.2...0.3.0].
 [0.11.0...0.12.0]: https://github.com/weph/obsidian-bookshelf/compare/0.11.0...0.12.0
 [0.12.0...0.13.0]: https://github.com/weph/obsidian-bookshelf/compare/0.12.0...0.13.0
 [0.13.0...0.13.1]: https://github.com/weph/obsidian-bookshelf/compare/0.13.0...0.13.1
-[0.13.1...main]: https://github.com/weph/obsidian-bookshelf/compare/0.12.0...main
+[0.13.1...0.14.0]: https://github.com/weph/obsidian-bookshelf/compare/0.13.1...0.14.0
+[0.14.0...main]: https://github.com/weph/obsidian-bookshelf/compare/0.14.0...main